refactor to clean up script
